kommentek időhöz kötött törlésére backend oldali ellenőrzés, felhaználók csoportokhoz hozzáadására és törlésére logika és adatbázis hívások

+++ routes/comments.js
@@ -1,6 +1,6 @@
 import express from "express";
 import { createComment, getComments } from "../database.js";
-import { findUserByName } from "../helpers.js";
+import { findUserByName, findUserByEmail } from "../helpers.js";
 import * as queries from "../database.js";
 
 const router = express.Router();
@@ -35,11 +35,26 @@ router.post("/", async (req, res) => {
 router.delete("/", async (req, res) => {
   if (req.isAuthenticated()) {
     const { id } = req.query;
+    const user = await findUserByEmail(req.session.passport.user);
+
+    const comment = await queries.getCommentById(id);
+    if (!comment) {
+      return res.status(404).json({ error: "Komment nem található" });
+    }
+
+    const createdAt = new Date(comment.createdAt);
+    const now = new Date();
+    if (now - createdAt > 60 * 1000 && user.role !== "ADMIN") {
+      return res
+        .status(403)
+        .json({ error: "1 perc után már nem törölhető a beküldött komment!" });
+    }
+
     const posts = await queries.deleteComment(id);
     if (posts) {
       res.status(204).send();
     } else {
-      res.status(500).send("Adatbázis hiba történt!");
+      res.status(500).send("Adatbázis hiba történt a komment törlésekor!");
     }
   } else {
     return res.status(401).send({ msg: "Sikertelen azonosítás!" });
